Push to Docker Hub
This tries to push to both GHCR and Docker Hub.

For this to work the secrets DOCKERHUB_USERNAME and DOCKERHUB_TOKEN must
be set to a user name and personal access token from the docker hub
respectively in the Github repository configuration.

The docker hub namespace will default to the same name as the Github
repository owner. If this is not the case, the env entry
DOCKERHUB_NAMESPACE has to be adopted accordingly.